--- src/main/java/org/winwin/request/Greeting.java
@@ -1,20 +1,0 @@
-package org.winwin.request;
-
-public class Greeting {
-
-    private final long id;
-    private final String content;
-
-    public Greeting(long id, String content) {
-        this.id = id;
-        this.content = content;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public String getContent() {
-        return content;
-    }
-}
