+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 ## Spring DATASOURCE (DataSourceAutoConfiguration & DataSourceProperties)
-spring.datasource.url=jdbc:postgresql://localhost:5432/winwin
-spring.datasource.username= postgres
-spring.datasource.password= admin
+spring.datasource.url=jdbc:postgresql://localhost:5432/winwindb
+spring.datasource.username= winwindbuser
+spring.datasource.password= winwindbpassword
 
 # The SQL dialect makes Hibernate generate better SQL for the chosen database
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
